chore: refactor system controls from `Settings` to new component `SystemControls`
